+++ CHANGELOG.md
@@ -13,6 +13,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Added the `Graph.Edges` method for obtaining all edges as a slice.
 * Added the `Graph.UpdateEdge` method for updating the properties of an edge.
 * Added the `Store.UpdateEdge` method for updating the properties of an edge.
+* Added the `NewLike` functions for creating a new graph that is "like" the given graph.
 * Added the `EdgeAttributes` functional option for setting an entire edge attributes map.
 
 ## [0.19.0] - 2023-04-23